requirements.txt
pillow
tensorflow-gpu==2.2.0rc3
